@@ -20,19 +20,20 @@
     "release:cli": "node ./scripts/release.js",
     "lint": "eslint . --fix",
     "type-check": "vue-tsc --build --force",
-    "postinstall": "simple-git-hooks"
+    "postinstall": "simple-git-hooks",
+    "inspector": "pnpx node-modules-inspector"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "5.0.0",
-    "@types/node": "^22.13.14",
+    "@antfu/eslint-config": "5.2.2",
+    "@types/node": "^24.3.1",
     "cross-env": "^10.0.0",
-    "eslint": "^9.32.0",
+    "eslint": "^9.35.0",
     "eslint-plugin-format": "^1.0.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "3.3.3",
-    "shx": "^0.3.4",
+    "prettier": "3.6.2",
+    "shx": "^0.4.0",
     "simple-git-hooks": "^2.12.1",
-    "typescript": "~5.6.2"
+    "typescript": "~5.9.2"
   },
   "simple-git-hooks": {
     "pre-commit": "npx lint-staged"
