@@ -30,7 +30,7 @@ const delegatedProps = computed(() => {
     <span
       v-if="props.label"
       :class="cn('text-xs text-muted-foreground bg-background absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 flex justify-center items-center',
-                 props.orientation === 'vertical' ? 'w-[1px] px-1 py-2' : 'h-[1px] py-1 px-2',
+                 props.orientation === 'vertical' ? 'w-px px-1 py-2' : 'h-px py-1 px-2',
       )"
     >{{ props.label }}</span>
   </Separator>
